feat(molecules): integrate EmailStep with atom components and move it to molecules folder
